refactor(dpop): rename sendTokenRequest to TokenService, preserve error shape
- Rename shared/openId4VCI/sendTokenRequest.(ts|test.ts) to TokenService
  per review feedback to match the codebase's *Service.ts naming
  convention (e.g. IssuersService.ts, backupService.ts).
- toVciClientError() now also sets code/message alongside
  issuerErrorCode/issuerErrorMessage so the thrown VciClientErrorResponse
  keeps the shape VciClient.ts already expects elsewhere.
- Add a test covering a generateTokenDPoPProof() rejection during the
  use_dpop_nonce retry, and assert code/message on the non-retry error
  test.
